+++ core/dbt/config/runtime.py
@@ -328,10 +328,19 @@ class RuntimeConfig(Project, Profile, AdapterRequiredConfig):
         if self.dependencies is None:
             all_projects = {self.project_name: self}
             internal_packages = get_include_paths(self.credentials.type)
-            project_paths = itertools.chain(
-                internal_packages,
-                self._get_project_directories()
-            )
+            # raise exception if fewer installed packages than in packages.yml
+            package_dirs = self._get_project_directories()
+            count_packages_specified = len(self.packages.packages)
+            count_packages_installed = len(tuple(package_dirs))
+            if count_packages_specified > count_packages_installed:
+                raise_compiler_error(
+                    f'dbt found {count_packages_specified} package(s) '
+                    f'specified in packages.yml, but only '
+                    f'{count_packages_installed} package(s) installed '
+                    f'in {self.modules_path}. Run "dbt deps" to '
+                    f'install package dependencies.'
+                )
+            project_paths = itertools.chain(internal_packages, package_dirs)
             for project_name, project in self.load_projects(project_paths):
                 if project_name in all_projects:
                     raise_compiler_error(