+++ core/dbt/exceptions.py
@@ -293,9 +293,9 @@ class UndefinedMacroException(CompilationException):
 
     def __str__(self, prefix='! ') -> str:
         msg = super().__str__(prefix)
-        return f'{msg}. This can happen when calling a macro that is ' \
-            'missing. Do you need to install a package dependency with '\
-            '`dbt deps`?'
+        return f'{msg}. This can happen when calling a macro that does ' \
+            'not exist. Check for typos and/or install package dependencies ' \
+            'with "dbt deps".'
 
 
 class UnknownAsyncIDException(Exception):